pyproject.toml
[project]
name = "devmate"
version = "0.1.0"
description = "AI-powered coding assistant with MCP, RAG, and Skills"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-mcp-adapters>=0.2",
    "langsmith",
    "chromadb",
    "tavily-python",
    "mcp[cli]",
    "starlette",
    "uvicorn[standard]",
    "click",
    "toml",
    "langchain-text-splitters",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/devmate", "mcp_server"]

[project.scripts]
devmate = "devmate.__main__:cli"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio",
    "ruff>=0.6",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio",
    "ruff>=0.6",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "W", "F", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
